importing, data cleaning, and creatingSchema utilities created - ha
